POC: Simplified app setup by moving the new_app.py to frigate/api
